+++ invenio/modules/oaiharvester/workflows/oaiharvest_harvest_repositories.py
@@ -1,5 +1,5 @@
 ## This file is part of Invenio.
-## Copyright (C) 2013, 2014 CERN.
+## Copyright (C) 2014 CERN.
 ##
 ## Invenio is free software; you can redistribute it and/or
 ## modify it under the terms of the GNU General Public License as
@@ -15,82 +15,55 @@
 ## along with Invenio; if not, write to the Free Software Foundation, Inc.,
 ## 59 Temple Place, Suite 330, Boston, MA 02111 1307, USA.
 
-"""Implements a typical ingestion workflow for MARCXML records in BibSched."""
+"""Main workflow iterating over selected repositories and downloaded files."""
 
-from ..tasks.marcxml_tasks import (get_repositories_list, harvest_records,
-                                   init_harvesting, get_records_from_file,
-                                   get_obj_extra_data_key, update_last_update,
-                                   filtering_oai_pmh_identifier)
+from invenio.modules.workflows.tasks.marcxml_tasks import (
+    get_obj_extra_data_key,
+    update_last_update,
+)
 
-from ..tasks.workflows_tasks import (start_workflow, workflows_reviews,
-                                     wait_for_a_workflow_to_complete,
-                                     get_nb_workflow_created,
-                                     get_workflows_progress,
-                                     write_something_generic,
-                                     num_workflow_running_greater)
+from invenio.modules.workflows.tasks.workflows_tasks import (
+    start_workflow,
+    workflows_reviews,
+    wait_for_a_workflow_to_complete,
+    get_nb_workflow_created,
+    get_workflows_progress,
+    write_something_generic,
+    num_workflow_running_greater
+)
 
-from ..tasks.logic_tasks import (foreach, end_for, simple_for, workflow_if,
-                                 workflow_else)
+from invenio.modules.workflows.tasks.logic_tasks import (
+    foreach,
+    end_for,
+    simple_for,
+    workflow_if,
+    workflow_else
+)
 
-from invenio.legacy.bibsched.bibtask import task_update_progress, write_message
+from invenio.legacy.bibsched.bibtask import (
+    task_update_progress,
+    write_message
+)
+from invenio.modules.workflows.definitions import WorkflowBase
 
-from invenio.modules.workflows.utils import WorkflowBase
+
+from ..tasks.harvesting import (
+    filtering_oai_pmh_identifier,
+    init_harvesting,
+    get_records_from_file,
+    get_repositories_list,
+    harvest_records,
+)
 
 
-class generic_harvesting_workflow_with_bibsched(WorkflowBase):
+class oaiharvest_harvest_repositories(WorkflowBase):
 
-    """A generic harvesting workflow for use with OAI harvesting in BibSched."""
+    """A workflow for use with OAI harvesting in BibSched."""
 
     object_type = "workflow"
 
-    @staticmethod
-    def get_description(bwo):
-        """Return a human readable description for data in workflow."""
-        from flask import render_template
-
-        identifiers = None
-
-        extra_data = bwo.get_extra_data()
-        try:
-            if 'options' in extra_data and 'identifiers' in extra_data["options"]:
-                identifiers = extra_data["options"]["identifiers"]
-
-            results = bwo.get_tasks_results()
-            result_progress = {}
-
-            if 'review_workflow' in results:
-                result_progress = results['review_workflow'][0]['result']
-            elif 'wait_for_a_workflow_to_complete' in results:
-                result_progress = results['wait_for_a_workflow_to_complete'][0]['result']
-
-            current_task = extra_data['_last_task_name']
-        except Exception as e:
-            result_progress = {}
-            identifiers = None
-            from invenio.modules.workflows.models import ObjectVersion
-            if bwo.version == ObjectVersion.INITIAL:
-                current_task = "The process has not started!!!"
-            else:
-                current_task = "The process CRASHED!!! \n {0}".format(str(e.message))
-
-        return render_template("workflows/styles/harvesting_description.html",
-                               identifiers=identifiers,
-                               result_progress=result_progress,
-                               current_task=current_task)
-
-    @staticmethod
-    def get_title(bwo):
-        """Return a human readable title for data in workflow."""
-        return "Supervising harvesting of {0}".format(
-            bwo.get_extra_data()["_repository"]["name"])
-
-    @staticmethod
-    def formatter(bwo, **kwargs):
-        """Return a human readable representation of data in workflow."""
-        return generic_harvesting_workflow_with_bibsched.get_description(bwo)
-
     workflow = [
-        write_something_generic("Initialisation", [task_update_progress,
+        write_something_generic("Initialization", [task_update_progress,
                                                    write_message]),
         init_harvesting,
         write_something_generic("Starting", [task_update_progress,
@@ -104,7 +77,7 @@ class generic_harvesting_workflow_with_bibsched(WorkflowBase):
                                                       write_message]),
             foreach(get_obj_extra_data_key("harvested_files_list")),
             [
-                write_something_generic("Creating Workflows",
+                write_something_generic("Creating workflows",
                                         [task_update_progress, write_message]),
                 foreach(get_records_from_file()),
                 [
@@ -112,7 +85,7 @@ class generic_harvesting_workflow_with_bibsched(WorkflowBase):
                     [
                         workflow_if(num_workflow_running_greater(10), neg=True),
                         [
-                            start_workflow("full_doc_process"),
+                            start_workflow("oaiharvest_record_post_process"),
 
                             write_something_generic(
                                 ["Workflow started: ",
@@ -128,7 +101,7 @@ class generic_harvesting_workflow_with_bibsched(WorkflowBase):
                                 [task_update_progress,
                                  write_message]),
                             wait_for_a_workflow_to_complete(0.05),
-                            start_workflow("full_doc_process", None),
+                            start_workflow("oaiharvest_record_post_process", None),
                             write_something_generic(["Workflow started :",
                                                      get_nb_workflow_created,
                                                      " "],
@@ -148,7 +121,7 @@ class generic_harvesting_workflow_with_bibsched(WorkflowBase):
         simple_for(0, get_nb_workflow_created, 1),
         [
             wait_for_a_workflow_to_complete(0.05),
-            write_something_generic([get_workflows_progress, " % Complete"],
+            write_something_generic([get_workflows_progress, " %% Complete"],
                                     [task_update_progress, write_message]),
         ],
         end_for,
