+++ invenio/modules/workflows/workflows/sample_workflow.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+#
 ## This file is part of Invenio.
 ## Copyright (C) 2012, 2013, 2014 CERN.
 ##
@@ -16,19 +17,17 @@
 ## along with Invenio; if not, write to the Free Software Foundation, Inc.,
 ## 59 Temple Place, Suite 330, Boston, MA 02111-1307, USA.
 
+"""Showcase a sample workflow definition."""
+
+from ..definitions import WorkflowBase
 from ..tasks.sample_tasks import add_data
 from ..tasks.logic_tasks import execute_if
 from ..tasks.marcxml_tasks import approve_record
-from invenio.modules.workflows.utils import WorkflowBase
 
 
 class sample_workflow(WorkflowBase):
-    """
-    This is a sample workflow.
-    """
-    @staticmethod
-    def formatter(bwo, **kwargs):
-        return ""
+
+    """This is a sample workflow."""
 
     workflow = [execute_if(add_data(1), lambda obj, eng: True),
                 approve_record]
